requirements.txt
fastapi
pydantic
uvicorn[standard]
asyncpg
python-multipart
SQLAlchemy
python-dotenv
pandas
openpyxl